sm-nist.h sm-nist.c - export sm_nist_decode_cvc and sm_nist_clear_cvc_content
nist-decode_cvc and nist_clear_cvc_content were renamed and exported via libopensc.exports
nist_cvc_t was moved to sm-nist.h

This allows them to be callable from outside of sm-nist to parse a CVC certificate.
To be added later is sm_nist_encode_cvc

 On branch myeid-4.9.1
 Changes to be committed:
	modified:   sm/sm-nist.c
	modified:   sm/sm-nist.h
